requirements.txt
boto3
python-dotenv
pyaudio
openai-whisper
pyautogui
